# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Technical Test Backend Aksamedia",
        "description": "API Documentation for Technical Test Backend Aksamedia",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://aksamedia-backed.laravel.cloud",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Ambil daftar pegawai",
                "operationId": "getEmployees",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter nama",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "division_id",
                        "in": "query",
                        "description": "Filter divisi ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "employees": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/EmployeeResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Tambah pegawai baru",
                "operationId": "storeEmployee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Detail pegawai",
                "operationId": "showEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EmployeeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update pegawai",
                "description": "Gunakan method POST dengan field '_method' = 'PUT' untuk update data yang mengandung file gambar.",
                "operationId": "updateEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmployeeResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Hapus pegawai",
                "operationId": "deleteEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil profil user yang sedang login",
                "description": "Mengembalikan detail data user berdasarkan token Bearer yang dikirim.",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data profil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User profile retrieved"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / Token Invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "EmployeeRequest": {
                "title": "Employee Request",
                "description": "Format input untuk create/update pegawai (Multipart/Form-Data)",
                "properties": {
                    "image": {
                        "description": "File foto pegawai (jpg, png, max 2MB)",
                        "type": "string",
                        "format": "binary"
                    },
                    "name": {
                        "description": "Nama Lengkap Pegawai",
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "phone": {
                        "description": "Nomor Telepon (Unik)",
                        "type": "string",
                        "example": "081234567890"
                    },
                    "division": {
                        "description": "UUID Divisi",
                        "type": "string",
                        "format": "uuid",
                        "example": "98a7s6d5-4f3g-2h1j-..."
                    },
                    "position": {
                        "description": "Jabatan",
                        "type": "string",
                        "example": "Staff IT"
                    },
                    "_method": {
                        "description": "Wajib diisi 'PUT' hanya saat update data dengan file gambar",
                        "type": "string",
                        "example": "PUT"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "User Resource",
                "description": "Schema data user/admin",
                "properties": {
                    "id": {
                        "description": "ID User",
                        "type": "string",
                        "format": "uuid",
                        "example": "98a7s6d5-..."
                    },
                    "name": {
                        "description": "Nama Lengkap",
                        "type": "string",
                        "example": "Administrator"
                    },
                    "username": {
                        "description": "Username Login",
                        "type": "string",
                        "example": "admin"
                    },
                    "phone": {
                        "description": "Nomor Telepon",
                        "type": "string",
                        "example": "08123456789"
                    },
                    "email": {
                        "description": "Email User",
                        "type": "string",
                        "format": "email",
                        "example": "admin@aksamedia.com"
                    }
                },
                "type": "object"
            },
            "DivisionResource": {
                "title": "Division Resource",
                "description": "Schema data untuk object Divisi",
                "properties": {
                    "id": {
                        "description": "ID Divisi",
                        "type": "string",
                        "format": "uuid",
                        "example": "98a7s6d5-4f3g-2h1j-..."
                    },
                    "name": {
                        "description": "Nama Divisi",
                        "type": "string",
                        "example": "Mobile Development"
                    }
                },
                "type": "object"
            },
            "EmployeeResource": {
                "title": "Employee Resource",
                "description": "Schema output data pegawai",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "uuid-1234"
                    },
                    "image": {
                        "type": "string",
                        "example": "http://domain.com/storage/uploads/foto.jpg"
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "phone": {
                        "type": "string",
                        "example": "08123456789"
                    },
                    "division": {
                        "$ref": "#/components/schemas/DivisionResource"
                    },
                    "position": {
                        "type": "string",
                        "example": "Staff IT"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}